Some more bug fixes
- Added invalidate call to mErrorLabel if exception occurs. Ran into an
issue and didn't see this updating. Not sure if this will solve it or
not but added it.
- Fixed issue where 3% max rise in price from buy order was working on
orders when we had open sell orders. Only want that when there are no
open sell orders.
- Put a stop-gap fix in for when we restart and don't have last
fulfilled sell information, but app expects it. Replaced with current
bid price if mLastFulfilledSell == null